+++ NEWS.md
@@ -14,6 +14,7 @@ devtools::install_github('EricArcher/rfPermute')
 * Fixed bug in plot.rfPermute that was reporting the p-value incorrectly at the top of the figure.
 * Fixed multi-threading in rfPermute so it works on Windows too.
 * Added impHeatmap function.
+* Switched proximity.plot to use ggplot2 graphics.
 
 ## Changes in 2.0: